refactor: share one segment walk between the artifact and knowledge guards
Both services had grown their own copy of the same primitive: canonicalise a
path one segment at a time so a symlinked ancestor stays usable while an
intermediate hop the caller never named is refused. Move the single
implementation to services/canonical-path.ts.

The knowledge importer names directories outside any managed tree, so it passes
no boundary and only the filesystem-root platform alias is relaxed — exactly
what its own walk did. The merged root-alias relaxation keeps the knowledge
walk's "must land inside the same filesystem root" qualifier, which is
unconditionally true on POSIX.

world-artifact-service re-exports resolveCanonicalPathWithoutSymlinkHops and
SymlinkHopError so its import surface is unchanged, and every call site keeps
its own error code and message.
